trading_bot/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
joblib==1.4.2
nltk==3.9.1
numpy==2.2.4
openai==1.75.0
pandas==2.2.3
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
textblob==0.19.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
